requirements.txt
pandas==2.3.3
openpyxl==3.1.5
matplotlib==3.10.8
textstat==0.7.13
spacy==3.8.13
rouge-score==0.1.2
streamlit==1.55.0
PyYAML==6.0.2
